pyproject.toml
[tool.poetry]
name = "ml_server"
version = "0.1.0"
description = "Multitask ML server"
authors = ["you@example.com"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = "^0.30.0"
python-dotenv = "^1.0.1"
scikit-learn = "^1.4.2"
joblib = "^1.4.2"
pydantic = "^2.6.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
